+++ rs/embedders/benches/compilation.rs
@@ -1,6 +1,7 @@
+use candid::Encode;
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 
-use embedders_bench::PostSetupAction;
+use embedders_bench::SetupAction;
 use ic_config::embedders::Config as EmbeddersConfig;
 use ic_embedders::{
     wasm_utils::{compile, validate_and_instrument_for_testing},
@@ -102,6 +103,12 @@ fn generate_binaries() -> Vec<(String, BinaryEncodedWasm)> {
     let qrcode_wasm = BinaryEncodedWasm::new(buf);
 
     result.push(("qrcode".to_string(), qrcode_wasm));
+
+    // This benchmark uses a canister from the motoko playground which is stored
+    // as a binary file in this repo.  It is generated from
+    // https://github.com/dfinity/motoko-playground/tree/abk/for-replica-benchmarking
+    let motoko_wasm = BinaryEncodedWasm::new(include_bytes!("test-data/pool.wasm").to_vec());
+    result.push(("motoko".to_string(), motoko_wasm));
 
     result
 }
@@ -250,11 +257,11 @@ fn execution(c: &mut Criterion) {
             c,
             &name,
             wasm.as_slice(),
-            &[],
+            &Encode!(&()).unwrap(),
             "go",
-            &[],
+            &Encode!(&()).unwrap(),
             None,
-            PostSetupAction::None,
+            SetupAction::None,
         );
     }
 }