+++ rs/ledger_suite/common/ledger_canister_core/src/ledger.rs
@@ -1,6 +1,6 @@
 use crate::{
     archive::ArchiveCanisterWasm,
-    blockchain::{BlockData, Blockchain},
+    blockchain::{BlockDataContainer, Blockchain},
     range_utils,
     runtime::Runtime,
 };
@@ -146,7 +146,7 @@ pub trait LedgerData: LedgerContext {
     type Transaction: LedgerTransaction<AccountId = Self::AccountId, Tokens = Self::Tokens>
         + Ord
         + Clone;
-    type BlockData: BlockData + Serialize + Default + for<'a> Deserialize<'a>;
+    type BlockDataContainer: BlockDataContainer + Default;
 
     // Purge configuration
 
@@ -170,10 +170,11 @@ pub trait LedgerData: LedgerContext {
 
     // Ledger data structures
 
-    fn blockchain(&self) -> &Blockchain<Self::Runtime, Self::ArchiveWasm, Self::BlockData>;
+    fn blockchain(&self)
+        -> &Blockchain<Self::Runtime, Self::ArchiveWasm, Self::BlockDataContainer>;
     fn blockchain_mut(
         &mut self,
-    ) -> &mut Blockchain<Self::Runtime, Self::ArchiveWasm, Self::BlockData>;
+    ) -> &mut Blockchain<Self::Runtime, Self::ArchiveWasm, Self::BlockDataContainer>;
 
     fn transactions_by_hash(&self) -> &BTreeMap<HashOf<Self::Transaction>, BlockIndex>;
     fn transactions_by_hash_mut(&mut self) -> &mut BTreeMap<HashOf<Self::Transaction>, BlockIndex>;